Break out stream code and remodel to look-ahead instead of backtracking
